+++ pybop/models/base_model.py
@@ -1,10 +1,12 @@
 from __future__ import annotations
+
+import copy
 from dataclasses import dataclass
 from typing import Any, Dict, Optional
-import pybamm
-import copy
-import numpy as np
+
 import casadi
+import numpy as np
+import pybamm
 
 Inputs = Dict[str, float]
 
@@ -12,7 +14,7 @@ Inputs = Dict[str, float]
 @dataclass
 class TimeSeriesState(object):
     """
-    The current state of a time series model that is a pybamm model
+    The current state of a time series model that is a pybamm model.
     """
 
     sol: pybamm.Solution
@@ -44,7 +46,7 @@ class BaseModel:
 
     """
 
-    def __init__(self, name="Base Model"):
+    def __init__(self, name="Base Model", parameter_set=None):
         """
         Initialize the BaseModel with an optional name.
 
@@ -54,10 +56,20 @@ class BaseModel:
             The name given to the model instance.
         """
         self.name = name
+        if parameter_set is None:
+            self._parameter_set = None
+        elif isinstance(parameter_set, dict):
+            self._parameter_set = pybamm.ParameterValues(parameter_set)
+        elif isinstance(parameter_set, pybamm.ParameterValues):
+            self._parameter_set = parameter_set
+        else:  # a pybop parameter set
+            self._parameter_set = pybamm.ParameterValues(parameter_set.params)
+
         self.pybamm_model = None
         self.parameters = None
         self.dataset = None
         self.signal = None
+        self.additional_variables = []
         self.matched_parameters = {}
         self.non_matched_parameters = {}
         self.fit_keys = []
@@ -343,15 +355,22 @@ class BaseModel:
                     inputs=inputs,
                     allow_infeasible_solutions=self.allow_infeasible_solutions,
                 ):
-                    sol = self.solver.solve(
-                        self.built_model, inputs=inputs, t_eval=t_eval
-                    )
+                    try:
+                        sol = self.solver.solve(
+                            self.built_model, inputs=inputs, t_eval=t_eval
+                        )
+                    except Exception as e:
+                        print(f"Error: {e}")
+                        return [np.inf]
                 else:
-                    return [np.inf]
+                    return {signal: [np.inf] for signal in self.signal}
 
-            simulation = [sol[signal].data for signal in self.signal]
+            y = {
+                signal: sol[signal].data
+                for signal in (self.signal + self.additional_variables)
+            }
 
-            return np.vstack(simulation).T
+            return y
 
     def simulateS1(self, inputs, t_eval):
         """
@@ -392,26 +411,40 @@ class BaseModel:
                 inputs=inputs,
                 allow_infeasible_solutions=self.allow_infeasible_solutions,
             ):
-                sol = self._solver.solve(
-                    self.built_model,
-                    inputs=inputs,
-                    t_eval=t_eval,
-                    calculate_sensitivities=True,
-                )
+                try:
+                    sol = self._solver.solve(
+                        self.built_model,
+                        inputs=inputs,
+                        t_eval=t_eval,
+                        calculate_sensitivities=True,
+                    )
+                    y = {signal: sol[signal].data for signal in self.signal}
 
-                simulation = [sol[signal].data for signal in self.signal]
+                    # Extract the sensitivities and stack them along a new axis for each signal
+                    dy = np.empty(
+                        (
+                            sol[self.signal[0]].data.shape[0],
+                            self.n_outputs,
+                            self.n_parameters,
+                        )
+                    )
 
-                sensitivities = [
-                    np.array(
-                        [[sol[signal].sensitivities[key]] for signal in self.signal]
-                    ).reshape(len(sol[self.signal[0]].data), self.n_outputs)
-                    for key in self.fit_keys
-                ]
+                    for i, signal in enumerate(self.signal):
+                        dy[:, i, :] = np.stack(
+                            [
+                                sol[signal].sensitivities[key].toarray()[:, 0]
+                                for key in self.fit_keys
+                            ],
+                            axis=-1,
+                        )
 
-                return np.vstack(simulation).T, np.dstack(sensitivities)
+                    return y, dy
+                except Exception as e:
+                    print(f"Error: {e}")
+                    return [np.inf], [np.inf]
 
             else:
-                return [np.inf], [np.inf]
+                return {signal: [np.inf] for signal in self.signal}, [np.inf]
 
     def predict(
         self,