+++ pybop/models/empirical/__init__.py
@@ -1,4 +1,5 @@
 #
 # Import lithium ion based models
 #
+from .base_ecm import ECircuitModel
 from .ecm import Thevenin