+++ pybop/models/empirical/ecm.py
@@ -1,5 +1,6 @@
 import pybamm
-from .ecm_base import ECircuitModel
+
+from .base_ecm import ECircuitModel
 
 
 class Thevenin(ECircuitModel):
@@ -44,24 +45,21 @@ class Thevenin(ECircuitModel):
         options=None,
         **kwargs,
     ):
-        super().__init__()
+        super().__init__(name, parameter_set)
         self.pybamm_model = pybamm.equivalent_circuit.Thevenin(
             options=options, **kwargs
         )
         self._unprocessed_model = self.pybamm_model
-        self.name = name
 
-        if isinstance(parameter_set, dict):
-            self.default_parameter_values = pybamm.ParameterValues(parameter_set)
-            self._parameter_set = self.default_parameter_values
+        # Set parameters, using either the provided ones or the default
+        if isinstance(self._parameter_set, dict):
+            self.default_parameter_values = pybamm.ParameterValues(self._parameter_set)
         else:
             self.default_parameter_values = self.pybamm_model.default_parameter_values
-            self._parameter_set = (
-                parameter_set or self.pybamm_model.default_parameter_values
-            )
-
+        self._parameter_set = self._parameter_set or self.default_parameter_values
         self._unprocessed_parameter_set = self._parameter_set
 
+        # Define model geometry and discretization
         self.geometry = geometry or self.pybamm_model.default_geometry
         self.submesh_types = submesh_types or self.pybamm_model.default_submesh_types
         self.var_pts = var_pts or self.pybamm_model.default_var_pts
@@ -70,6 +68,7 @@ class Thevenin(ECircuitModel):
         )
         self.solver = solver or self.pybamm_model.default_solver
 
+        # Internal attributes for the built model are initialized but not set
         self._model_with_set_params = None
         self._built_model = None
         self._built_initial_soc = None