+++ tests/unit/test_dataset.py
@@ -1,6 +1,7 @@
-import pytest
-import pybop
 import numpy as np
+import pytest
+
+import pybop
 
 
 class TestDataset:
@@ -39,3 +40,15 @@ class TestDataset:
         # Test conversion of pybamm solution into dictionary
         assert dataset.data == pybop.Dataset(solution).data
         assert dataset.names == pybop.Dataset(solution).names
+
+        # Test set and get item
+        test_current = solution["Current [A]"].data + np.ones_like(
+            solution["Current [A]"].data
+        )
+        dataset["Current [A]"] = test_current
+        assert np.all(dataset["Current [A]"] == test_current)
+        with pytest.raises(ValueError):
+            dataset["Time"]
+
+        # Test conversion of single signal to list
+        assert dataset.check(signal="Terminal voltage [V]")