+++ tests/unit/test_standalone.py
@@ -1,6 +1,7 @@
-import pytest
-import pybop
 import numpy as np
+import pytest
+
+import pybop
 from examples.standalone.cost import StandaloneCost
 from examples.standalone.problem import StandaloneProblem
 
@@ -53,13 +54,18 @@ class TestStandalone:
 
         # Test the Problem with a Cost
         rmse_cost = pybop.RootMeanSquaredError(problem)
-        x = rmse_cost([1, 2])
+        rmse_x = rmse_cost([1, 2])
+        rmse_grad_x = rmse_cost.evaluateS1([1, 2])
 
-        np.testing.assert_allclose(x, 3.138, atol=1e-2)
+        np.testing.assert_allclose(rmse_x, 3.05615, atol=1e-2)
+        np.testing.assert_allclose(rmse_grad_x[1], [-0.81758337, 0.0], atol=1e-2)
 
         # Test the sensitivities
         sums_cost = pybop.SumSquaredError(problem)
-        sums_cost.evaluateS1([1, 2])
+        x = sums_cost.evaluateS1([1, 2])
+
+        np.testing.assert_allclose(x[0], 934.006734006734, atol=1e-2)
+        np.testing.assert_allclose(x[1], [-334.006734, 0.0], atol=1e-2)
 
         # Test incorrect number of initial parameter values
         with pytest.raises(ValueError):